LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/013-expected.txt


FAIL Link with onclick navigation to javascript url with delayed document.write and href navigation  assert_array_equals: property 0, expected "after script" but got "javascript"
 Test
